refactor(qmia): merge benchmark scripts into single benchmark_qmia.py
