+++ examples/sklearn/summarize_qmia_lira_benchmark.py
@@ -168,7 +168,7 @@ def parse_args() -> argparse.Namespace:
         "benchmark_json",
         type=str,
         nargs="+",
-        help="One or more JSON files generated by benchmark_qmia_vs_lira.py.",
+        help="One or more JSON files generated by benchmark_qmia.py.",
     )
     return parser.parse_args()
